chore: refactor to share contributor type filter
