+++ components/molecules/MostActiveContributorsCard/most-active-contributors-card.tsx
@@ -3,20 +3,11 @@ import { useGesture } from "@use-gesture/react";
 import Image from "next/image";
 import { ReactNode, useState } from "react";
 import * as RawTooltip from "@radix-ui/react-tooltip";
-import Button from "components/atoms/Button/button";
 import Card from "components/atoms/Card/card";
-import Icon from "components/atoms/Icon/icon";
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuTrigger,
-} from "components/atoms/Dropdown/dropdown";
+
 import { getAvatarByUsername } from "lib/utils/github";
-import PeopleIcon from "img/icons/people.svg";
-import ChevronDownIcon from "img/chevron-down.svg";
-import SVGIcon from "components/atoms/SVGIcon/svg-icon";
 import Tooltip from "components/atoms/Tooltip/tooltip";
+import { ContributorType, ContributorTypeFilter } from "components/Graphs/shared/contributor-type-filter";
 
 // omit total_contributions and login from ContributorStat
 type StatKeys = keyof Omit<ContributorStat, "total_contributions" | "login">;
@@ -40,8 +31,6 @@ export interface ContributorStat {
   total_contributions: number;
 }
 
-export type ContributorType = "all" | "active" | "new" | "alumni";
-
 interface Props {
   topContributor?: ContributorStat;
   data: ContributorStat[];
@@ -50,13 +39,6 @@ interface Props {
   isLoading: boolean;
   totalContributions: number;
 }
-
-const peopleFilters: Record<ContributorType, string> = {
-  all: "All Contributors",
-  active: "Active Contributors",
-  new: "New Contributors",
-  alumni: "Alumni Contributors",
-};
 
 const LegendItem = ({ color, title }: { color?: string; title: string }) => {
   return (
@@ -136,26 +118,7 @@ export default function MostActiveContributorsCard({
 
         {/* buttons */}
         <div className="flex gap-1 mb-4">
-          <DropdownMenu>
-            <DropdownMenuTrigger asChild>
-              <Button variant="default" className="items-center gap-1">
-                <SVGIcon IconImage={`${PeopleIcon.src}#icon`} className="w-4 h-4" />
-                {peopleFilters[contributorType]}
-                <Icon IconImage={ChevronDownIcon} className="w-4 h-4" />
-              </Button>
-            </DropdownMenuTrigger>
-            <DropdownMenuContent align="end" className="flex flex-col gap-2">
-              {Object.entries(peopleFilters).map(([key, value]) => (
-                <DropdownMenuItem
-                  key={key}
-                  className="rounded-md !cursor-pointer"
-                  onClick={() => setContributorType(key as keyof typeof peopleFilters)}
-                >
-                  {value}
-                </DropdownMenuItem>
-              ))}
-            </DropdownMenuContent>
-          </DropdownMenu>
+          <ContributorTypeFilter contributorType={contributorType} setContributorType={setContributorType} />
         </div>
         {/* chart */}
         <div className="relative grid place-content-stretch overflow-y-hidden mb-4">
